+++ test/core/camera_test.cpp
@@ -1,6 +1,7 @@
 #include "core/camera.h"
-#include "core/vec3.h"
 #include "core/ray.h"
+#include "core/vec3.h"
+
 #include <gtest/gtest.h>
 
 TEST(CameraTest, CameraDeclaration) {
