+++ test/core/colour_test.cpp
@@ -1,5 +1,6 @@
 #include "core/colour.h"
 #include "core/vec3.h"
+
 #include <gtest/gtest.h>
 
 TEST(ColourTest, ScaleColour) {
