+++ test/core/image_test.cpp
@@ -1,5 +1,6 @@
 #include "core/image.h"
 #include "core/vec3.h"
+
 #include <gtest/gtest.h>
 
 TEST(ImageTest, ImageDeclaration) {
